Consolidate C++ code into a single file
After recent changes we have just few lines of code spread across
few files, so consolidate everything into a single file.
